Refactoring NavBar Components
