+++ frontend/src/views/Dashboard.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import Layout from '../components/Layout/Layout';
 import TestDataVisualization from './TestDataVisualization.js';
-import Navbar from '../components/Layout/Navbar';
+import Navbar from '../components/Layout/NavBar/Navbar';
 
 const Dashboard = props => {
   return (
